+++ tests/testthat/test_4_textPlot.R
@@ -24,7 +24,7 @@ test_that("textProjection MEAN and PCA produces a tibble with character variable
 
   expect_true(tibble::is_tibble(df_for_plotting))
   expect_is(df_for_plotting$words[1], "character")
-  expect_is(df_for_plotting$n[1], "integer")
+  expect_is(df_for_plotting$n[1], "numeric")
 })
 
 test_that("textProjection with QUARTILE, 0.9 PCA and 2-dimensions produces a tibble with character variable and numeric variable.", {
@@ -52,7 +52,7 @@ test_that("textProjection with QUARTILE, 0.9 PCA and 2-dimensions produces a tib
 
   expect_true(tibble::is_tibble(df_for_plotting))
   expect_is(df_for_plotting$words[1], "character")
-  expect_is(df_for_plotting$n[1], "integer")
+  expect_is(df_for_plotting$n[1], "numeric")
 })
 
 test_that("textProjectionPlot 1-DIMENSIONS produces a plot", {
@@ -75,7 +75,10 @@ test_that("textProjectionPlot 1-DIMENSIONS produces a plot", {
     x_axes_label = "Low vs. High HILS score",
     y_axes_label = "Low vs. High SWLS score",
     p_adjust_method = "bonferroni",
-    scale_y_axes_lim = NULL
+    scale_y_axes_lim = NULL,
+    group_embeddings1 = F,
+    group_embeddings2 = F,
+    projection_embedding = F,
   )
 
   expect_true(ggplot2::is.ggplot(p$final_plot))
@@ -102,7 +105,10 @@ test_that("textProjectionPlot 1-DIMENSIONS produces a plot", {
     x_axes_label = "Low vs. High HILS score",
     y_axes_label = "Low vs. High SWLS score",
     p_adjust_method = "bonferroni",
-    scale_y_axes_lim = NULL
+    scale_y_axes_lim = NULL,
+    group_embeddings1 = T,
+    group_embeddings2 = T,
+    projection_embedding = T,
   )
 
   expect_true(ggplot2::is.ggplot(p$final_plot))